pr review changes part 4 of n, remove original kWhrs variable to use battery params more directly, add comments on grid point variables, move shared_rate_data to correct folder
