+++ shared/lib_battery_dispatch_automatic_btm.cpp
@@ -854,7 +854,7 @@ void dispatch_automatic_behind_the_meter_t::costToCycle()
     {
         if (curr_year < m_battReplacementCostPerKWH.size()) {
             double capacityPercentDamagePerCycle = _Battery->estimateCycleDamage();
-            m_cycleCost = 0.01 * capacityPercentDamagePerCycle * m_battReplacementCostPerKWH[curr_year] * m_battOriginalKWH;
+            m_cycleCost = 0.01 * capacityPercentDamagePerCycle * m_battReplacementCostPerKWH[curr_year] * _Battery->get_params().nominal_energy;
         }
         else {
             // Should only apply to BattWatts. BattWatts doesn't have price signal dispatch, so this is fine.
@@ -863,6 +863,6 @@ void dispatch_automatic_behind_the_meter_t::costToCycle()
     }
     else if (m_battCycleCostChoice == dispatch_t::INPUT_CYCLE_COST)
     {
-        m_cycleCost = cycle_costs_by_year[curr_year] * m_battOriginalKWH;
+        m_cycleCost = cycle_costs_by_year[curr_year] * _Battery->get_params().nominal_energy;
     }
 }
