+++ test/input_cases/shared_rate_data.cpp
